ml/stt-requirements.txt
# Requirements for Speech-to-Text with Audio Chunking
# Use: pip install -r sst-requirements.txt

# Audio processing
librosa>=0.11.0
pydub>=0.25.1
soundfile>=0.13.1

# Machine learning & numerical processing
numpy>=2.1.0
torch>=2.6.0

# Speech recognition
openai-whisper>=20240900  # OpenAI's Whisper model

# Utilities
tqdm>=4.67.0
ffmpeg-python>=0.2.0  # For audio processing support 